# backend/docs/swagger-docs.yaml
openapi: 3.0.1
info:
  title: Alfa Backend API
  description: API для управления клиентами и документами.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Локальный сервер

paths:
  /client:
    get:
      summary: Получение клиента по ID
      description: Возвращает объект клиента на основе идентификатора клиента.
      parameters:
        - name: clientId
          in: query
          required: true
          description: Идентификатор клиента
          schema:
            type: integer
      responses:
        '200':
          description: Клиент успешно найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '404':
          description: Клиент не найден

  /predict:
    get:
      summary: Предсказание метода подписи
      description: Возвращает предсказанный метод подписи для указанного клиента.
      parameters:
        - name: clientId
          in: query
          required: true
          description: Идентификатор клиента
          schema:
            type: integer
      responses:
        '200':
          description: Предсказание успешно выполнено
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Клиент не найден

  /current-method:
    get:
      summary: Получение текущего метода подписи
      description: Возвращает текущий метод подписи для клиента.
      parameters:
        - name: clientId
          in: query
          required: true
          description: Идентификатор клиента
          schema:
            type: integer
      responses:
        '200':
          description: Успешно получен текущий метод подписи
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Клиент не найден

  /claims-count:
    get:
      summary: Получение количества жалоб клиента
      description: Возвращает количество жалоб (claims) клиента.
      parameters:
        - name: clientId
          in: query
          required: true
          description: Идентификатор клиента
          schema:
            type: integer
      responses:
        '200':
          description: Количество жалоб успешно получено
          content:
            application/json:
              schema:
                type: integer
        '404':
          description: Клиент не найден

  /allowed-signatures:
    get:
      summary: Получение доступных методов подписи
      description: Возвращает список доступных методов подписи на основе наличия мобильного приложения у клиента.
      parameters:
        - name: clientId
          in: query
          required: true
          description: Идентификатор клиента
          schema:
            type: integer
      responses:
        '200':
          description: Доступные методы успешно получены
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '404':
          description: Клиент не найден

  /predict-on-document:
    post:
      summary: Предсказание на основе документа
      description: Возвращает предсказанный метод подписи на основе переданных данных.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PredictionRequestPayload'
      responses:
        '200':
          description: Предсказание успешно выполнено
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Неверные данные запроса

  /change-signature:
    patch:
      summary: Смена подписи используемой клиентом
      description: Изменяет текущий метод подписи клиента.
      parameters:
        - name: clientId
          in: query
          required: true
          description: Идентификатор клиента
          schema:
            type: integer
        - name: signature
          in: query
          required: true
          description: Новый метод подписи
          schema:
            type: string
            enum: [SMS, EMAIL, QES]
      responses:
        '200':
          description: Подпись успешно изменена
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Ошибка

  /signed-docs:
    get:
      summary: Получение подписанных документов клиента
      description: Возвращает список всех подписанных документов для указанного клиента.
      parameters:
        - name: clientId
          in: query
          required: true
          description: Идентификатор клиента
          schema:
            type: integer
      responses:
        '200':
          description: Список подписанных документов успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
        '404':
          description: Клиент не найден или документы отсутствуют
        '400':
          description: Некорректный запрос, отсутствует clientId

  /sign-document:
    patch:
      summary: Подписание документа
      description: Позволяет подписать документ, устанавливая флаг isSigned в true.
      parameters:
        - name: documentId
          in: query
          required: true
          description: Идентификатор документа, который нужно подписать
          schema:
            type: integer
      responses:
        '200':
          description: Документ успешно подписан
        '404':
          description: Документ с указанным ID не найден

components:
  schemas:
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        currentMethod:
          type: string
          enum: [SMS, EMAIL, QES]
        mobileApp:
          type: boolean
        claims:
          type: integer
        signedDocumentsBasicMobile:
          type: integer
        signedDocumentsBasicWeb:
          type: integer
        signedDocumentsImportantMobile:
          type: integer
        signedDocumentsImportantWeb:
          type: integer

    PredictionRequestPayload:
      type: object
      properties:
        clientId:
          type: integer
        documentId:
          type: integer
        platformType:
          type: string

    Document:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        isSigned:
          type: boolean